Fix potential lag from natural armor repair menu
The current implementation of the "repair armor" menu option
eagerly scans every item on the map when the menu is displayed,
which can cause lag. As a fix, use GenClosest for each ingredient def
with a custom validator to ensure only the relevant items are scanned,
and reorder the checks to avoid doing this at all if the armor of
the pawn being inspected cannot be repaired any further.