Consistently apply shot height adjustment
0f52a9fa8aedae8c969a399a75d91fde38d6d708 introduced a system to adjust
the shot height dynamically to compensate for the case where the target
is partially obscured by cover or is tall enough for its center of mass
to be higher than the shot height. However, this is currently applied
only inconsistently: while all projectiles take it into account
for shot line checks, only instant (IE: laser) weapons actually
apply it to the projectile being fired.

Fix it by making the height and angle adjustments part of ShiftVecReport
and using them as needed.