Remove LINQ from suppression code
Suppression code can be fairly hot, e.g. when an explosion occurs
within a large group of pawns that sends them looking for cover.
Some of the relevant logic in SuppressionUtility is using LINQ
heavily, which causes significant memory pressure in such scenarios.
When profiling a save with a doomsday hitting a 10k point tribal raid,
InterceptorZonesFor() was responsible for ~3.7% of total time
despite there being no shields on the map at all.

So, replace LINQ in relevant methods with iterator functions,
which are less memory intensive, eliminating about ~3.5% of the total time.