Remove legacy JobGiver_ConfigurableHostilityResponse
This was replaced by a Harmony patch in 2019 (7ea3841), but was later inadvertently added back to compilation after the csproj was overhauled. That in turn broke the Harmony patch because it started targeting the legacy class rather than RimWorld's own vanilla JobGiver_ConfigurableHostilityResponse, causing red errors when a pawn set to attack nearby hostiles automatically had an empty weapon and encountered a hostile pawn.