Update JI weapons
Moved defese type ammos to the core mod, as they now rely on the same ammo. Projecitle graphic is patched by sec mod.
Increased the AMR's range by 10, 68 is a bit short for the weapon with bipod deployed given its role.
Changed the bases on a few ammo types to remove trails.