Let manned turrets shoot over adjacent pawns
To avoid friendly fire, pawns only shoot over allies if they are
adjacent to them. However, manned turrets calculate adjacency from the
position of the shooter, rather than the turret itself, which yields
incorrect results. So, consider the position of the caster, be it pawn
or turret, rather than the position of the shooting pawn.