Fix angle and velocity calculation for vehicle mortars
Mortar shells use a custom `gravityFactor` defined in projectile props
that vehicle turret shot angle calculation logic currently doesn't take into account. So,
inject it by way of a transpiler since the current callback doesn't give
us access to the projectile def being used. Also divide the default
value by MetersPerCellWidth because we're operating on cells here.

For this to work, we also need to consult CompCharges to dynamically
adjust the projectile speed based on gravity and shot range, and feed
the adjusted speed into both the angle calculation and the launch
proper.