Add test project and workflow for LoadFolders.xml validation
We already have a Python script that checks for duplicate LoadFolders.xml
entries, but it hasn't been updated for 1.6 and doesn't currently run in CI.
Convert it into a .NET SDK test project and add a workflow to run it for PRs.
Have it also check that the directories defined in LoadFolders.xml actually exist.